+++ cryptofeed/backends/backend.py
@@ -50,17 +50,23 @@ class BackendCallback:
 
 
 class BackendBookCallback:
+    async def _write_snapshot(self, book, receipt_timestamp: float):
+        data = book.to_dict(as_type=self.numeric_type)
+        del data['delta']
+        data['receipt_timestamp'] = receipt_timestamp
+        await self.write(data)
+
     async def __call__(self, book, receipt_timestamp: float):
-        if self.snapshots_only or self.snapshot_interval == self.snapshot_count[book.symbol]:
-            data = book.to_dict(as_type=self.numeric_type)
-            del data['delta']
-            data['receipt_timestamp'] = receipt_timestamp
-            await self.write(data)
-            self.snapshot_count[book.symbol] = 0
+        if self.snapshots_only:
+            await self._write_snapshot(book, receipt_timestamp)
         else:
             data = book.to_dict(delta=book.delta is not None, as_type=self.numeric_type)
             data['receipt_timestamp'] = receipt_timestamp
             if book.delta is None:
                 del data['delta']
+            else:
+                self.snapshot_count[book.symbol] += 1
             await self.write(data)
-            self.snapshot_count[book.symbol] += 1
+            if self.snapshot_interval <= self.snapshot_count[book.symbol] and book.delta:
+                await self._write_snapshot(book, receipt_timestamp)
+                self.snapshot_count[book.symbol] = 0