+++ cryptofeed/exchanges/binance.py
@@ -324,6 +324,7 @@ class Binance(Feed, BinanceRestMixin):
 
         url = f'{self.rest_endpoint}/depth?symbol={pair}&limit={max_depth}'
         resp = await self.http_conn.read(url)
+        await sleep(1 / self.request_limit)
         resp = json.loads(resp, parse_float=Decimal)
         return resp
 
@@ -479,7 +480,6 @@ class Binance(Feed, BinanceRestMixin):
             ]
         }
         """
-        book_args = (msg, pair, timestamp)
         std_pair = self.exchange_symbol_to_std_symbol(pair)
         is_expired_snapshot = self.refresh_snapshot and std_pair in self._next_snapshot_time and time.time() > self._next_snapshot_time[std_pair]
 
@@ -490,11 +490,11 @@ class Binance(Feed, BinanceRestMixin):
             elif is_expired_snapshot:
                 # Refresh snapshot to make sure our data is still up to date
                 await self._refresh_snapshot(pair)
-            return await self._handle_book_msg(*book_args) 
+            return await self._handle_book_msg(msg, pair, timestamp) 
 
         if std_pair in self._book_buffer:
             # snapshot is currently being fetched. std_pair will exist in self._l2_book after snapshot, but we need to continue buffering until all previous buffered messages have been processed.
-            return self._book_buffer[std_pair].append(book_args)
+            return self._book_buffer[std_pair].append((msg, pair, timestamp))
 
         if is_expired_snapshot:
             # Refresh snapshot to make sure our data is still up to date
@@ -503,15 +503,16 @@ class Binance(Feed, BinanceRestMixin):
                 self._fetch_snapshop_buffer[std_pair] = deque()
                 create_task(self._refresh_snapshot_with_buffer(pair))
             self._fetch_snapshop_buffer[std_pair].append(msg)
-            return await self._handle_book_msg(*book_args)
+            return await self._handle_book_msg(msg, pair, timestamp)
 
         elif std_pair in self._l2_book:
             # snapshot exists
-            return await self._handle_book_msg(*book_args)
+            return await self._handle_book_msg(msg, pair, timestamp)
 
-        # Initiate buffer & get snapshot
+        # concurrent http enabled and no buffer exists, so
+        # initiate buffer & get snapshot
         self._book_buffer[std_pair] = deque()
-        self._book_buffer[std_pair].append(book_args)
+        self._book_buffer[std_pair].append((msg, pair, timestamp))
 
         async def _concurrent_snapshot():
             await self._reset_snapshot(pair)